Rename package so it does not conflict with readability-lxml which is also just named readability.
