+++ example.py
@@ -1,15 +1,18 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
-from readability import Reader
+from readability_selenium import Reader
 
 from selenium import webdriver
 
 import sys
 
+options = webdriver.ChromeOptions()
+options.add_argument('--headless')
+options.headless = True
+
 driver = webdriver.Remote(
   command_executor='http://localhost:4444/wd/hub',
-  desired_capabilities={'browserName': 'firefox', 'javascriptEnabled': True})
-
+  desired_capabilities=options.to_capabilities())
 
 # to supply a full path to Readability.js, do this
 #js = open("/tmp/Readability.js").read()
@@ -18,7 +21,8 @@ driver = webdriver.Remote(
 r = Reader(driver)
 
 url = sys.argv[1]
-a = r.get_url(url)
+a = r.get_readable_dict(url)
 
-print(a)
+print(a['content'])
+print(a['byline'])
